feat(agents): Implement redirects for agent presets and add approvals page
- Added a redirect for agent presets to streamline navigation.
- Created a new approvals page to manage agent approvals.
- Updated existing components to reflect the new routing structure and feature flags.
